src/api/requirements.txt
fastapi==0.111.1
pyjwt==2.8.0